requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
langchain-openai==0.0.2
chromadb
pypdf
streamlit
pandas
requests
fastapi 
uvicorn
pysqlite3-binary ; sys_platform == 'linux'